+++ tests/workers/emptyWorker.js
@@ -1,5 +1,5 @@
 'use strict'
-const { ThreadWorker } = require('../../lib/workers')
+const { ThreadWorker } = require('../../lib/index')
 
 function test (data) {}
